% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{calc_pos}
\alias{calc_pos}
\title{calculates the weighted mean point between two points}
\usage{
calc_pos(x0, x1, pos = 0.5)
}
\arguments{
\item{x0}{first point}

\item{x1}{second point}

\item{pos}{weighting towards first point, defaults to 0.5}
}
\description{
calculates the weighted mean point between two points
}
