% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{find_centre}
\alias{find_centre}
\title{finds centre of a flodia object}
\usage{
find_centre(object)
}
\arguments{
\item{object}{list containing at list x0, x1, y0 and y1}
}
\description{
finds centre of a flodia object
}
