% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flodia}
\alias{flodia}
\title{plot a flodia}
\usage{
flodia(f, oma = rep(0.1, 4), args = list())
}
\arguments{
\item{f}{flodia plot function}

\item{oma}{numeric of length 4 specifying outer margin of plot, at x0, y0,
x1, and y1. Default is 0.1}

\item{args}{list of named arguments to \code{f}}
}
\description{
plot a flodia
}
