% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flodia_null}
\alias{flodia_null}
\title{run a flodia without plotting (e.g to extract co-ordinates)}
\usage{
flodia_null(f, args = list(), oma = rep(0.1, 4))
}
\arguments{
\item{f}{flodia plot function}

\item{args}{list of named arguments to \code{f}}

\item{oma}{numeric of length 4 specifying outer margin of plot, at x0, y0,
x1, and y1. Default is 0.1}
}
\description{
run a flodia without plotting (e.g to extract co-ordinates)
}
