% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flodia_pdf}
\alias{flodia_pdf}
\title{save a flodia as pdf}
\usage{
flodia_pdf(f, filepath, width = 19/2.5, oma = rep(0.1, 4), args = list(), ...)
}
\arguments{
\item{f}{flodia plot function}

\item{filepath}{desired location of output eg "test.png"}

\item{width}{width in pixels of output (height is calculated automatically
based on plot dimensions)}

\item{oma}{numeric of length 4 specifying outer margin of plot, at x0, y0,
x1, and y1. Default is 0.1}

\item{args}{list of named arguments to \code{f}}

\item{...}{other arguments to pdf()}
}
\description{
save a flodia as pdf
}
