% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{flow}
\alias{flow}
\title{add a flow between two nodes}
\usage{
flow(
  node0 = NULL,
  node1 = NULL,
  label = "",
  pos = 0.5,
  label_pos = 0.5,
  label_gap = 0.1,
  label_col = "black",
  font = 3,
  label_cex = 0.8,
  arr_type = "triangle",
  arr_length = 0.2,
  arr_width = 0.15,
  ...
)
}
\arguments{
\item{node0}{start node}

\item{node1}{end node}

\item{label}{label to draw on flow}

\item{pos}{a decimal between 0 and 1 giving the position to depart from the
start node, where 0 = bottom / left and 1 = top/right}

\item{label_pos}{a decimal between 0 and 1 giving the position along the flow
to draw the label, where 0 = bottom / left and 1 = top/right}

\item{label_gap}{distance from the flow at which to draw the label,
default = 0.1}

\item{label_col}{colour of label, defaults to black}

\item{font}{font of label, defaults to 3}

\item{label_cex}{label cex, defualts to 0.8}

\item{arr_type}{type of arrow head to draw default = triangle}

\item{arr_length}{length of arrow head, defaults to 0.2}

\item{arr_width}{width of arrow head, defaults to 0.15}

\item{...}{additional formatting arguments to arrow()}
}
\value{
returns the start and end points of the flow
}
\description{
add a flow between two nodes
}
