% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fork.R
\name{forky}
\alias{forky}
\title{add fork between three nodes, in the y-direction}
\usage{
forky(
  from,
  to0,
  to1,
  pos = NULL,
  label_from = NULL,
  label_to0 = NULL,
  label_to1 = NULL,
  pos_from = NULL,
  pos_to0 = NULL,
  pos_to1 = pos_to0,
  label_from_pos = NULL,
  label_to0_pos = label_from_pos,
  label_to1_pos = label_to0_pos,
  label_from_gap = NULL,
  label_to0_gap = label_from_gap,
  label_to1_gap = label_to0_gap,
  label_from_x = NULL,
  label_to0_x = NULL,
  label_to1_x = NULL,
  label_from_y = NULL,
  label_to0_y = NULL,
  label_to1_y = label_to0_y,
  arr_width = NULL,
  name_from = deparse(substitute(from)),
  name_to0 = deparse(substitute(to0)),
  name_to1 = deparse(substitute(to1)),
  ...
)
}
\arguments{
\item{from}{start node}

\item{to0}{the left end node}

\item{to1}{the right end node}

\item{pos}{a decimal between 0 and 1 giving the position between \code{from}, and
the closest of \code{to0} and \code{to1} at which to split the fork.}

\item{label_from}{label to draw on the flow out of \code{from}}

\item{label_to0}{label to draw on the flow into \code{to0}}

\item{label_to1}{label to draw on the flow into \code{to1}}

\item{pos_from}{a decimal between 0 and 1 giving the position to start the
flow out of \code{from}, where 0 = left and 1 = right}

\item{pos_to0}{a decimal between 0 and 1 giving the position to attach the
flow to node \code{to0}, where 0 = left and 1 = right}

\item{pos_to1}{a decimal between 0 and 1 giving the position to attach the
flow to node \code{to1}, where 0 = left and 1 = right}

\item{label_from_pos}{a decimal between 0 and 1 giving the position along the
flow out of \code{from} to draw \code{label_from}, where 0 = start and 1 = end}

\item{label_to0_pos}{a decimal between 0 and 1 giving the position along the
flow into \code{to0} to draw \code{label_to0}, where 0 = start and 1 = end}

\item{label_to1_pos}{a decimal between 0 and 1 giving the position along the
flow into \code{to1} to draw \code{label_to1}, where 0 = start and 1 = end}

\item{label_from_gap}{distance from the flow out of \code{from} at which to draw
\code{label_from} default = 0.05}

\item{label_to0_gap}{distance from the flow into \code{to0} at which to draw
\code{label_to0} default = 0.05}

\item{label_to1_gap}{distance from the flow into \code{to1} at which to draw
\code{label_to1} default = 0.05}

\item{label_from_x}{x co-ordinate of \code{label_from} position, overrides use of
\code{label_from_pos} and / or \code{label_from_gap}}

\item{label_to0_x}{x co-ordinate of \code{label_to0} position, overrides use of
\code{label_to0_pos} and / or \code{label_to0_gap}}

\item{label_to1_x}{x co-ordinate of \code{label_to1} position, overrides use of
\code{label_to1_pos} and / or \code{label_to1_gap}}

\item{label_from_y}{y co-ordinate of \code{label_from} position, overrides use of
\code{label_from_pos} and / or \code{label_from_gap}}

\item{label_to0_y}{y co-ordinate of \code{label_to0} position, overrides use of
\code{label_to0_pos} and / or \code{label_to0_gap}}

\item{label_to1_y}{y co-ordinate of \code{label_to1} position, overrides use of
\code{label_to1_pos} and / or \code{label_to1_gap}}

\item{arr_width}{width of arrow, defaults to same as \code{\link[=flow]{flow()}}}

\item{name_from}{internal argument used for informative error messages}

\item{name_to0}{internal argument used for informative error messages}

\item{name_to1}{internal argument used for informative error messages}

\item{...}{additional formatting arguments to \code{\link[=flow]{flow()}}}
}
\value{
returns the start and end points of the flow
}
\description{
add fork between three nodes, in the y-direction
}
