% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{group a set of nodes and flows}
\usage{
group(x0, y0, x1, y1, gap = rep(0.1, 4), col = "grey90", border = "black")
}
\arguments{
\item{x0}{left-most x}

\item{y0}{bottom-most y}

\item{x1}{right-most x}

\item{y1}{top-most y}

\item{gap}{required border around nodes, vector of length 4 specifying:
bottom, left, top, right. defaults to 0.1 on all sides}

\item{col}{colour of group box}

\item{border}{colour of group border}
}
\value{
coordinates of group
}
\description{
group a set of nodes and flows
}
