% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflow.R
\name{inflowx}
\alias{inflowx}
\title{add inflow to a node, in direction x}
\usage{
inflowx(to, length, pos = NULL, ...)
}
\arguments{
\item{to}{end node}

\item{length}{single numeric specifying length, positive will result in a
flow from left to right, negative from right to left}

\item{pos}{a decimal between 0 and 1 giving the position to attach the
flow to node \code{to}, where 0 = bottom and 1 = top, default = 0.5}

\item{...}{additional formatting arguments to \code{\link[=flow]{flow()}}}
}
\value{
returns the start and end points of the flow
}
\description{
add inflow to a node, in direction x
}
