% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{mid_palette}
\alias{mid_palette}
\title{mid palette}
\usage{
mid_palette(col)
}
\arguments{
\item{col}{either an integer vector between 1 and X or a string of the
following possible colour names: rd, or, oryl, yl, ylgn, gn, gnbu, bu, bupu,
pu, pupi, pi, gr}
}
\description{
mid palette
}
