% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node}
\alias{node}
\title{add a node}
\usage{
node(
  x,
  y,
  r,
  label = "",
  x_root = 0,
  y_root = 0,
  node_col = "grey80",
  border = "black",
  font = 1,
  label_col = "black"
)
}
\arguments{
\item{x}{cental x position}

\item{y}{central y position}

\item{r}{radius of node}

\item{label}{node label}

\item{x_root}{shift from x = 0 (default = 0)}

\item{y_root}{shift from y = 0 (default = 0)}

\item{node_col}{colour of node}

\item{border}{colour of node border}

\item{font}{label font, see par()}

\item{label_col}{colour of}
}
\description{
add a node
}
