% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outflow.R
\name{outflowx}
\alias{outflowx}
\title{add outflow to a node, in direction x}
\usage{
outflowx(from, length, pos = NULL, ...)
}
\arguments{
\item{from}{start node}

\item{length}{single numeric specifying length, positive will result in a
flow from left to right, negative from right to left}

\item{pos}{a decimal between 0 and 1 giving the position to start the
flow out of \code{from}, where 0 = bottom and 1 = top, default = 0.5}

\item{...}{additional formatting arguments to \code{\link[=flow]{flow()}}}
}
\value{
returns the start and end points of the flow
}
\description{
add outflow to a node, in direction x
}
