% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outflow.R
\name{outflowy}
\alias{outflowy}
\title{add outflow to a node, in direction y}
\usage{
outflowy(from, length, pos = NULL, ...)
}
\arguments{
\item{from}{end node}

\item{length}{single numeric specifying length, positive will result in a
flow from bottom to top, negative from top to bottom}

\item{pos}{a decimal between 0 and 1 giving the position to start the
flow out of \code{from}, where 0 = left and 1 = right, default = 0.5}

\item{...}{additional formatting arguments to \code{\link[=flow]{flow()}}}
}
\value{
returns the start and end points of the flow
}
\description{
add outflow to a node, in direction y
}
