% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitx}
\alias{splitx}
\title{add split between three nodes, in the x-direction}
\usage{
splitx(
  from = NULL,
  to0,
  to1,
  length = NULL,
  label_from = NULL,
  label_to0 = NULL,
  label_to1 = NULL,
  pos_from = NULL,
  pos_to = NULL,
  label_pos = NULL,
  label_from_pos = label_pos,
  label_to0_pos = NULL,
  label_to1_pos = NULL,
  label_gap = NULL,
  label_from_gap = label_gap,
  label_to0_gap = NULL,
  label_to1_gap = NULL,
  label_from_x = NULL,
  label_to0_x = NULL,
  label_to1_x = NULL,
  label_from_y = NULL,
  label_to0_y = NULL,
  label_to1_y = NULL,
  arr_width = NULL,
  arr_width0 = arr_width,
  arr_width1 = arr_width,
  name_from = deparse(substitute(from)),
  name_to0 = deparse(substitute(to0)),
  name_to1 = deparse(substitute(to1)),
  ...
)
}
\arguments{
\item{from}{start node}

\item{to0}{the bottom end node}

\item{to1}{the top end node}

\item{length}{single numeric specifying length, positive will result in a
fork from left to right, negative from right to left. Optional parameter
replacing \code{from}}

\item{label_from}{label to draw on the flow out of \code{from}}

\item{label_to0}{label to draw on the flow into \code{to0}}

\item{label_to1}{label to draw on the flow into \code{to1}}

\item{pos_from}{a decimal between 0 and 1 giving the position to start the
flow out of \code{from}, where 0 = bottom and 1 = top}

\item{pos_to}{a decimal between 0 and 1 giving the position to attach the
flow to nodes \code{to0} and \code{to1}, where 0 = left and 1 = right}

\item{label_pos}{a decimal between 0 and 1 giving the position along the
flow at which to draw the labels.}

\item{label_from_pos}{a decimal between 0 and 1 giving the position along the
flow out of \code{from} to draw \code{label_from}, where 0 = start and 1 = end}

\item{label_to0_pos}{a decimal between 0 and 1 giving the position along the
flow into \code{to0} to draw \code{label_to0}, where 0 = start and 1 = end}

\item{label_to1_pos}{a decimal between 0 and 1 giving the position along the
flow into \code{to1} to draw \code{label_to1}, where 0 = start and 1 = end}

\item{label_gap}{distance from the flows at which to draw
labels default = 0.05}

\item{label_from_gap}{distance from the flow out of \code{from} at which to draw
\code{label_from} default = 0.05}

\item{label_to0_gap}{distance from the flow into \code{to0} at which to draw
\code{label_to0} default = 0.05}

\item{label_to1_gap}{distance from the flow into \code{to1} at which to draw
\code{label_to1} default = 0.05}

\item{label_from_x}{x co-ordinate of \code{label_from} position, overrides use of
\code{label_from_pos} and / or \code{label_from_gap}}

\item{label_to0_x}{x co-ordinate of \code{label_to0} position, overrides use of
\code{label_to0_pos} and / or \code{label_to0_gap}}

\item{label_to1_x}{x co-ordinate of \code{label_to1} position, overrides use of
\code{label_to1_pos} and / or \code{label_to1_gap}}

\item{label_from_y}{y co-ordinate of \code{label_from} position, overrides use of
\code{label_from_pos} and / or \code{label_from_gap}}

\item{label_to0_y}{y co-ordinate of \code{label_to0} position, overrides use of
\code{label_to0_pos} and / or \code{label_to0_gap}}

\item{label_to1_y}{y co-ordinate of \code{label_to1} position, overrides use of
\code{label_to1_pos} and / or \code{label_to1_gap}}

\item{arr_width}{width of arrows, defaults to same as \code{\link[=flow]{flow()}}}

\item{arr_width0}{width of arrow into \code{to0}, defaults to same as \code{\link[=flow]{flow()}}}

\item{arr_width1}{width of arrow into \code{to1}, defaults to same as \code{\link[=flow]{flow()}}}

\item{name_from}{internal argument used for informative error messages}

\item{name_to0}{internal argument used for informative error messages}

\item{name_to1}{internal argument used for informative error messages}

\item{...}{additional formatting arguments to \code{\link[=flow]{flow()}}}
}
\value{
returns the start and end points of the flow
}
\description{
add split between three nodes, in the x-direction
}
