% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splity}
\alias{splity}
\title{add split between three nodes, in direction y}
\usage{
splity(
  from,
  to0,
  to1,
  label_from = NULL,
  label_to0 = NULL,
  label_to1 = NULL,
  pos_from = NULL,
  pos_to = NULL,
  label_from_pos = NULL,
  label_to0_pos = NULL,
  label_to1_pos = NULL,
  label_from_gap = NULL,
  label_to0_gap = NULL,
  label_to1_gap = NULL,
  arr_width = NULL,
  ...
)
}
\arguments{
\item{from}{start node}

\item{to0}{the left end node}

\item{to1}{the right end node}

\item{label_from}{label to draw on the flow out of \code{from}}

\item{label_to0}{label to draw on the flow into \code{to0}}

\item{label_to1}{label to draw on the flow into \code{to1}}

\item{pos_from}{a decimal between 0 and 1 giving the position to start the
flow out of \code{from}, where 0 = left and 1 = right}

\item{pos_to}{a decimal between 0 and 1 giving the position to attach the
flow to nodes \code{to0} and \code{to1}, where 0 = left and 1 = right}

\item{label_from_pos}{a decimal between 0 and 1 giving the position along the
flow out of \code{from} to draw \code{label_from}, where 0 = start and 1 = end}

\item{label_to0_pos}{a decimal between 0 and 1 giving the position along the
flow into \code{to0} to draw \code{label_to0}, where 0 = start and 1 = end}

\item{label_to1_pos}{a decimal between 0 and 1 giving the position along the
flow into \code{to1} to draw \code{label_to1}, where 0 = start and 1 = end}

\item{label_from_gap}{distance from the flow out of \code{from} at which to draw
\code{label_from} default = 0.05}

\item{label_to0_gap}{distance from the flow into \code{to0} at which to draw
\code{label_to0} default = 0.05}

\item{label_to1_gap}{distance from the flow into \code{to1} at which to draw
\code{label_to1} default = 0.05}

\item{arr_width}{width of arrow, defaults to same as \code{\link[=flow]{flow()}}}

\item{...}{additional formatting arguments to \code{\link[=flow]{flow()}}}
}
\value{
returns the start and end points of the flow
}
\description{
add split between three nodes, in direction y
}
