% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{start_plot}
\alias{start_plot}
\title{start a flodia plot}
\usage{
start_plot(
  x0,
  y0,
  x1,
  y1,
  filepath,
  width = 1200,
  res = 200,
  oma = c(1, 1, 1, 1)
)
}
\arguments{
\item{x0}{left-most x}

\item{y0}{bottom-most y}

\item{x1}{right-most x}

\item{y1}{top-most y}

\item{filepath}{desired location of output eg "test.png"}

\item{width}{width in pixels of output (height is calculated automatically
based on plot dimensions)}

\item{res}{resolution of plot default = 200 dpi}

\item{oma}{outer margin of plot, default is 1}
}
\description{
start a flodia plot
}
