% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn.R
\name{turnx}
\alias{turnx}
\title{add turn between two nodes in the x-direction}
\usage{
turnx(
  from,
  mid,
  to,
  label_from = NULL,
  label_mid = NULL,
  label_to = NULL,
  pos_from = NULL,
  pos_to = NULL,
  label_from_pos = NULL,
  label_mid_pos = NULL,
  label_to_pos = NULL,
  label_from_gap = NULL,
  label_mid_gap = NULL,
  label_to_gap = NULL,
  label_from_x = NULL,
  label_mid_x = NULL,
  label_to_x = NULL,
  label_from_y = NULL,
  label_mid_y = NULL,
  label_to_y = NULL,
  arr_width = NULL,
  name_from = deparse(substitute(from)),
  name_to = deparse(substitute(to)),
  ...
)
}
\arguments{
\item{from}{start node}

\item{mid}{the y co-ordinate of the turn}

\item{to}{the end node}

\item{label_from}{label to draw on the flow out of \code{from}}

\item{label_mid}{label to draw on the \code{mid} flow}

\item{label_to}{label to draw on the flow into \code{to}}

\item{pos_from}{a decimal between 0 and 1 giving the position to start the
flow out of \code{from}, where 0 = left and 1 = right}

\item{pos_to}{a decimal between 0 and 1 giving the position to attach the
flow to node \code{to}, where 0 = left and 1 = right}

\item{label_from_pos}{a decimal between 0 and 1 giving the position along the
flow out of \code{from} to draw \code{label_from}, where 0 = start and 1 = end}

\item{label_mid_pos}{a decimal between 0 and 1 giving the position along the
\code{mid} flow to draw \code{label_mid}, where 0 = start and 1 = end}

\item{label_to_pos}{a decimal between 0 and 1 giving the position along the
flow into \code{to} to draw \code{label_to}, where 0 = start and 1 = end}

\item{label_from_gap}{distance from the flow out of \code{from} at which to draw
\code{label_from} default = 0.05}

\item{label_mid_gap}{distance from the \code{mid} flow at which to draw
\code{label_mid} default = 0.05}

\item{label_to_gap}{distance from the flow into \code{to} at which to draw
\code{label_to} default = 0.05}

\item{label_from_x}{x co-ordinate of \code{label_from} position, overrides use of
\code{label_from_pos} and / or \code{label_from_gap}}

\item{label_mid_x}{x co-ordinate of \code{label_mid} position, overrides use of
\code{label_mid_pos} and / or \code{label_mid_gap}}

\item{label_to_x}{x co-ordinate of \code{label_to} position, overrides use of
\code{label_to_pos} and / or \code{label_to_gap}}

\item{label_from_y}{y co-ordinate of \code{label_from} position, overrides use of
\code{label_from_pos} and / or \code{label_from_gap}}

\item{label_mid_y}{y co-ordinate of \code{label_mid} position, overrides use of
\code{label_mid_pos} and / or \code{label_mid_gap}}

\item{label_to_y}{y co-ordinate of \code{label_to} position, overrides use of
\code{label_to_pos} and / or \code{label_to_gap}}

\item{arr_width}{width of arrow, defaults to same as \code{\link[=flow]{flow()}}}

\item{name_from}{internal argument used for informative error messages}

\item{name_to}{internal argument used for informative error messages}

\item{...}{additional formatting arguments to \code{\link[=flow]{flow()}}}
}
\value{
returns the start and end points of the flow
}
\description{
add turn between two nodes in the x-direction
}
