% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn.R
\name{turny}
\alias{turny}
\title{add turn between two nodes in the y-direction}
\usage{
turny(
  from,
  mid_y,
  to,
  label_from = NULL,
  label = NULL,
  label_to = NULL,
  pos_from = NULL,
  pos_to = NULL,
  label_pos = NULL,
  label_gap = NULL,
  label_x = NULL,
  label_y = NULL,
  arr_width = NULL,
  name_from = deparse(substitute(from)),
  name_to = deparse(substitute(to)),
  ...
)
}
\arguments{
\item{from}{start node}

\item{mid_y}{the y co-ordinate of the mid flow}

\item{to}{the end node}

\item{label}{label to draw on the mid flow}

\item{pos_from}{a decimal between 0 and 1 giving the position to start the
flow out of \code{from}, where 0 = left and 1 = right}

\item{pos_to}{a decimal between 0 and 1 giving the position to attach the
flow to node \code{to}, where 0 = left and 1 = right}

\item{label_pos}{a decimal between 0 and 1 giving the position along the
\code{mid} flow to draw \code{label}, where 0 = left and 1 = right}

\item{label_gap}{distance from the \code{mid} flow at which to draw
\code{label} default = 0.05}

\item{label_x}{x co-ordinate of \code{label} position, overrides use of \code{label_pos}}

\item{label_y}{y co-ordinate of \code{label} position, overrides use of \code{label_gap}}

\item{arr_width}{width of arrow, defaults to same as \code{\link[=flow]{flow()}}}

\item{name_from}{internal argument used for informative error messages}

\item{name_to}{internal argument used for informative error messages}

\item{...}{additional formatting arguments to \code{\link[=flow]{flow()}}}
}
\value{
returns the start and end points of the flow
}
\description{
add turn between two nodes in the y-direction
}
