% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flodia}
\alias{flodia}
\title{plot a flodia}
\usage{
flodia(f, oma = 0.1)
}
\arguments{
\item{f}{flodia plot function}

\item{oma}{single numeric outer margin of plot, default is 1}
}
\description{
plot a flodia
}
