% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flodia_null}
\alias{flodia_null}
\title{run a flodia without plotting (e.g to extract co-ordinates)}
\usage{
flodia_null(f, oma = 0.1, args = list())
}
\arguments{
\item{f}{flodia plot function}

\item{oma}{single numeric outer margin of plot, default is 0.1}

\item{args}{list of named arguments to \code{f}}
}
\description{
run a flodia without plotting (e.g to extract co-ordinates)
}
