% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flodia_null}
\alias{flodia_null}
\title{run a flodia without plotting (e.g to extract co-ordinates)}
\usage{
flodia_null(f)
}
\arguments{
\item{f}{flodia plot function}
}
\description{
run a flodia without plotting (e.g to extract co-ordinates)
}
