% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flodia_png}
\alias{flodia_png}
\title{save a flodia}
\usage{
flodia_png(f, filepath, width = 1200, res = 200, oma = 0.1, args = list(), ...)
}
\arguments{
\item{f}{flodia plot function}

\item{filepath}{desired location of output eg "test.png"}

\item{width}{width in pixels of output (height is calculated automatically
based on plot dimensions)}

\item{res}{resolution of plot default = 200 dpi}

\item{oma}{single numeric outer margin of plot, default is 0.1}

\item{args}{list of named arguments to \code{f}}

\item{...}{other arguments to png()}
}
\description{
save a flodia
}
