% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node}
\alias{node}
\title{add a node}
\usage{
node(
  x,
  y,
  r = 0.1,
  label = "",
  rx = r,
  ry = r,
  x_root = 0,
  y_root = 0,
  node_col = "grey80",
  border = "black",
  label_font = 1,
  label_col = "black",
  label_cex = 1
)
}
\arguments{
\item{x}{cental x position}

\item{y}{central y position}

\item{r}{radius of node}

\item{label}{node label}

\item{rx}{radius of node in x-direction}

\item{ry}{radius of node in y-direction}

\item{x_root}{shift from x = 0 (default = 0)}

\item{y_root}{shift from y = 0 (default = 0)}

\item{node_col}{colour of node}

\item{border}{colour of node border}

\item{label_col}{colour of lavel}

\item{label_cex}{size of label text}

\item{font}{label font, see par()}
}
\description{
add a node
}
